Convert usercommands dictionaries into function with emitter as parameter so that the user do not publish global variables in module anymore
