+++ bindings/nodejs/lib/chatbot-api/chatbot-interface.js
@@ -39,113 +39,158 @@ CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
 var Q = require('q');
 
-var NO_CHATBOT_ERROR_MESSAGE = 'No model loaded in the chatbot';
+function treatMessage(chatbot, userCommandsDriver, sessionId, context, msg, deferred, saveContext) {
+    var currentState = context._state;
+    var intentFound = chatbot.treatMessage(context, msg, function (actionId, intentVariables) {
+        if (userCommandsDriver) {
+            userCommandsDriver.execute(actionId, sessionId, intentVariables)
+                .then(function (userDefinedVariables) {
+                    // Must create userDefinedVariables because it is mandatory in the C++ layer.
+                    if (!userDefinedVariables) {
+                        userDefinedVariables = {};
+                    }
 
-function ChatbotInterface(chatbot) {
-    this._chatbot = chatbot;
+                    //let the user override the branching completely, by specifying
+                    //the target state and a dynamic reply from the user_command
+                    var replies = [];
+                    if('_target_state' in userDefinedVariables && '_reply' in userDefinedVariables) {
+                        var targetState = userDefinedVariables['_target_state'];
+                        var reply = userDefinedVariables['_reply'];
 
-    this.talk = function(sessionId, message) {
-        var deferred = Q.defer();
+                        context._state = targetState;
+                        replies.push(reply);
+                    } else {
+                        replies = chatbot.prepareReplies(currentState, actionId, intentVariables, userDefinedVariables);
+                    }
 
-        if(!sessionId) {
-            deferred.reject('No sessionId provided');
-            return deferred.promise;
+                    saveContext(context);
+                    deferred.resolve(replies);
+                });
         }
-
-        if(!message) {
-            deferred.reject('No message provided');
-            return deferred.promise;
-        }
-
-        if(!this._chatbot) {
-            deferred.reject(NO_CHATBOT_ERROR_MESSAGE);
-            return deferred.promise;
-        }
-
-        this._chatbot.talk(sessionId, message)
-        .then(function(replies) {
+        else {
+            // No user defined variables.
+            var replies = chatbot.prepareReplies(currentState, actionId, intentVariables, {});
             deferred.resolve(replies);
-        })
-        .fail(function(err) {
-            deferred.reject(err);
-        });
-
-        return deferred.promise;
-    };
-
-    this.setState = function(sessionId, state) {
-        var deferred = Q.defer();
-
-        if(!sessionId) {
-            deferred.reject('No sessionId provided');
-            return deferred.promise;
         }
+    });
 
-        if(!state) {
-            deferred.reject('No state provided');
-            return deferred.promise;
-        }
-
-        if(!this._chatbot) {
-            deferred.reject(NO_CHATBOT_ERROR_MESSAGE);
-            return deferred.promise;
-        }
-
-        this._chatbot.setState(sessionId, state)
-        .then(function() {
-            deferred.resolve();
-        })
-        .fail(function(err) {
-            deferred.reject(err);
-        });
-
-        return deferred.promise;
-    };
-
-    this.getState = function(sessionId) {
-        var deferred = Q.defer();
-
-        if(!sessionId) {
-            deferred.reject('No sessionId provided');
-            return deferred.promise;
-        }
-
-        if(!this._chatbot) {
-            deferred.reject(NO_CHATBOT_ERROR_MESSAGE);
-            return deferred.promise;
-        }
-
-        this._chatbot.getState(sessionId)
-        .then(function(state) {
-            deferred.resolve(state);
-        })
-        .fail(function(err) {
-            deferred.reject(err);
-        });
-
-        return deferred.promise;
-    };
-
-    this.getGraph = function() {
-        var deferred = Q.defer();
-
-        if(!this._chatbot) {
-            deferred.reject(NO_CHATBOT_ERROR_MESSAGE);
-            return deferred.promise;
-        }
-
-        try {
-            var graph = this._chatbot.getGraph();
-            deferred.resolve(graph);
-        }
-        catch(err) {
-            deferred.reject();
-        }
-
-        return deferred.promise;
-    };
-
-    return this;
+    if(!intentFound) {
+        deferred.reject('No intent found');
+    }
 }
 
-module.exports = ChatbotInterface;
+function callTreatMessageWithContext(chatbot, sessionManager, userCommandsDriver, sessionId, context, msg, deferred) {
+    try {
+        var saveContext = (context) => (sessionManager.save(sessionId, context));
+        treatMessage(chatbot, userCommandsDriver, sessionId, context, msg, deferred, saveContext);
+    }
+    catch(err) {
+        console.log(err);
+        deferred.reject(err);
+    }
+}
+
+var talkInternal = function(chatbot, sessionManager, userCommandsDriver, sessionId, msg) {
+    var deferred = Q.defer();
+
+    sessionManager.load(sessionId)
+    .then(function(context) {
+        callTreatMessageWithContext(chatbot, sessionManager, userCommandsDriver, sessionId, context, msg, deferred);
+    })
+    // If the state does not exist in the session manager
+    .fail(function(err) {
+        var context = {
+            '_state': chatbot.getInitialState()
+        };
+        callTreatMessageWithContext(chatbot, sessionManager, userCommandsDriver, sessionId, context, msg, deferred);
+    });
+    return deferred.promise;
+};
+
+var setStateInternal = function(sessionManager, sessionId, state) {
+    var context = {
+        _state: state
+    };
+    return sessionManager.save(sessionId, context);
+};
+
+function getStateInternal(sessionManager, sessionId, initialState) {
+    var deferred = Q.defer();
+    sessionManager.load(sessionId)
+    .then(function(context) {
+        if('_state' in context) {
+            deferred.resolve(context._state);
+        }
+        else {
+            deferred.reject('No state in the context');
+        }
+    })
+    .fail(function() {
+        deferred.resolve(initialState);
+    });
+    return deferred.promise;
+}
+
+var getInitialStateInternal = function(chatbot) {
+    var defer = Q.defer();
+    defer.resolve(chatbot.getInitialState());
+    return defer.promise;
+};
+
+var getGraphInternal = function(chatbot) {
+    var defer = Q.defer();
+    defer.resolve(chatbot.getGraph(chatbot));
+    return defer.promise;
+};
+
+var setupUserCommands = function(userCommands) {
+    return userCommands.setup();
+};
+
+var cleanupUserCommands = function(userCommands) {
+    return userCommands.cleanup();
+};
+
+function OpenIntentChatbot(serializableChatbot, sessionManagerDriver, userCommandsDriver) {
+
+    if(!sessionManagerDriver) {
+        throw new Error('No session manager driver');
+    }
+
+    this._serializableChatbot = serializableChatbot;
+    this._userCommandsDriver = userCommandsDriver;
+    this._sessionManager = sessionManagerDriver;
+}
+
+OpenIntentChatbot.prototype.talk = function(sessionId, msg) {
+    return talkInternal(this._serializableChatbot, this._sessionManager, this._userCommandsDriver, sessionId, msg);
+};
+
+OpenIntentChatbot.prototype.setState = function(sessionId, state) {
+    return setStateInternal(this._sessionManager, sessionId, state);
+};
+
+OpenIntentChatbot.prototype.getState = function(sessionId) {
+    var initialState = this._serializableChatbot.getInitialState();
+    return getStateInternal(this._sessionManager, sessionId, initialState);
+};
+
+OpenIntentChatbot.prototype.getInitialState = function() {
+    return getInitialStateInternal(this._serializableChatbot);
+};
+
+OpenIntentChatbot.prototype.getGraph = function() {
+    return getGraphInternal(this._serializableChatbot);
+};
+
+OpenIntentChatbot.prototype.setupUserCommands = function() {
+    return setupUserCommands(this._userCommandsDriver);
+};
+
+OpenIntentChatbot.prototype.cleanupUserCommands = function() {
+    return cleanupUserCommands(this._userCommandsDriver);
+};
+
+
+
+module.exports = OpenIntentChatbot;
