+++ bindings/nodejs/project-skeletons/chatbot/res/user_commands.js
@@ -1,13 +1,14 @@
 var food_type;
 
-module.exports = {
-    "#get_food_type": function(intentVariables, sessionId, next) {
+module.exports = function(handler) {
+    handler.on("#get_food_type", function(intentVariables, sessionId, next) {
         var replyVariables = {};
         food_type = intentVariables['food_type0'];
         replyVariables['0'] = food_type;
         next(replyVariables);
-    },
-    "#confirm": function(intentVariables, sessionId, next) {
+    });
+
+    handler.on("#confirm", function(intentVariables, sessionId, next) {
         var replyVariables = {};
         if(food_type == 'pizza') {
             replyVariables['0'] = '8';
@@ -16,5 +17,5 @@ module.exports = {
             replyVariables['0'] = '5';
         }
         next(replyVariables);
-    }
-}
+    });
+};
