+++ bindings/nodejs/test/chatbot-api/res/food_bot/user_commands.js
@@ -37,14 +37,16 @@ AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY
 WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
-module.exports = {
-    "#get_food_type": function(intentVariables, sessionId, next) {
+module.exports = function(handler) {
+    
+    handler.on("#get_food_type", function(intentVariables, sessionId, next) {
         var replyVariables = {};
         replyVariables['0'] = intentVariables['food_type0'];
         next(replyVariables);
-    },
-    "#confirm": function(intentVariables, sessionId, next) {
+    });
+
+    handler.on("#confirm", function(intentVariables, sessionId, next) {
         replyVariables['0'] = '5';
         next(replyVariables);
-    }
+    });
 }
