+++ bindings/nodejs/test/res/food_bot/user_commands.js
@@ -37,8 +37,8 @@ AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY
 WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
-module.exports = {
-    "#get_food_type": function(intentVariables, sessionId, next) {
+module.exports = function(handler) {
+    handler.on("#get_food_type", function(intentVariables, sessionId, next) {
         var replyVariables = {};
         replyVariables['0'] = intentVariables['food_type0'];
 
@@ -49,10 +49,11 @@ module.exports = {
         }
 
         next(replyVariables);
-    },
-    "#confirm": function(intentVariables, sessionId, next) {
+    });
+
+    handler.on("#confirm", function(intentVariables, sessionId, next) {
         var replyVariables = {};
         replyVariables['0'] = '5';
         next(replyVariables);
-    }
-}
+    });
+};
